refactor: move all callback function to each pages
